abstract out @vltpkg/fast-split
This is useful in more places than just the semver parser.
