+++ src/fast-split/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@vltpkg/semver",
-  "description": "The semantic version parser used by vlt",
+  "name": "@vltpkg/fast-split",
+  "description": "A fast way to split small-to-medium sized strings by small string delimiters",
   "version": "0.0.0-0",
   "tshy": {
     "selfLink": false,
@@ -12,18 +12,12 @@
       ".": "./src/index.ts"
     }
   },
-  "dependencies": {
-    "@vltpkg/error-cause": "workspace:*",
-    "@vltpkg/fast-split": "workspace:*"
-  },
   "devDependencies": {
     "@eslint/js": "catalog:",
     "@types/eslint__js": "catalog:",
     "@types/node": "catalog:",
-    "@types/semver": "^7.5.8",
     "eslint": "catalog:",
     "prettier": "catalog:",
-    "semver": "^7.6.2",
     "tap": "catalog:",
     "tshy": "catalog:",
     "typedoc": "catalog:",
@@ -31,9 +25,6 @@
     "typescript-eslint": "catalog:"
   },
   "scripts": {
-    "benchmark": "./test/fixtures/vs-node-semver.ts",
-    "benchmark:bun": "hyperfine --warmup 3 'bun ./benchmarks/bun.js' 'bun ./benchmarks/npm.js' 'bun ./benchmarks/this.js'",
-    "benchmark:node": "hyperfine --warmup 3 'node ./benchmarks/npm.js' 'node ./benchmarks/this.js'",
     "format": "prettier --write . --log-level warn --ignore-path ../../.prettierignore --cache",
     "format:check": "prettier --check . --ignore-path ../../.prettierignore --cache",
     "lint": "eslint . --fix",
@@ -43,6 +34,7 @@
     "presnap": "npm run prepare",
     "snap": "tap",
     "pretest": "npm run prepare",
+    "benchmark:split": "./test/fixtures/split-vs-fast-split.ts",
     "test": "tap",
     "typedoc": "typedoc",
     "preversion": "npm test",
