+++ src/fast-split/test/fixtures/split-vs-fast-split.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S tap -Rtap --disable-coverage --allow-empty-coverage
 import t from 'tap'
-import { fastSplit } from '../../src/fast-split.js'
+import { fastSplit } from '../../src/index.js'
 
 const str = '1.2.3-asdf+foo'
