+++ src/fast-split/typedoc.json
@@ -1,0 +1,5 @@
+{
+  "entryPoints": ["./src/**/*.+(ts|tsx|mts|cts)"],
+  "extends": ["../../typedoc.base.json"],
+  "tsconfig": "./.tshy/esm.json"
+}
